Moved chickenKyiv functions to helper.js
